[Backport to 3.2.x][Fixes #7948] Error handling monitoring module (#7998)
(cherry picked from commit dea24e245b3dedc195a0b15913bbfd1268af9f58)

# Conflicts:
#	pavement.py
