+++ geonode/layers/urls.py
@@ -22,7 +22,7 @@ from geonode import geoserver  # noqa
 from geonode.utils import check_ogc_backend
 from django.conf.urls import url, include
 from django.views.generic import TemplateView
-from geonode.monitoring import register_url_event
+from geonode.base import register_url_event
 
 from . import views
