+++ geonode/upload/views.py
@@ -62,7 +62,7 @@ from geonode.base.enumerations import CHARSETS
 from geonode.utils import fixup_shp_columnnames
 from geonode.decorators import logged_in_or_basicauth
 
-from geonode.monitoring import register_event
+from geonode.base import register_event
 from geonode.monitoring.models import EventType
 
 from .forms import (
