+++ pavement.py
@@ -300,10 +300,12 @@ def static(options):
 def setup(options):
     """Get dependencies and prepare a GeoNode development environment."""
 
-    updategeoip(options)
-    info(('GeoNode development environment successfully set up.'
-          'If you have not set up an administrative account,'
-          ' please do so now. Use "paver start" to start up the server.'))
+    if MONITORING_ENABLED:
+        updategeoip(options)
+
+    info('GeoNode development environment successfully set up.'
+         'If you have not set up an administrative account,'
+         ' please do so now. Use "paver start" to start up the server.')
 
 
 def grab_winfiles(url, dest, packagename):
@@ -378,11 +380,12 @@ def updategeoip(options):
     """
     Update geoip db
     """
-    settings = options.get('settings', '')
-    if settings and 'DJANGO_SETTINGS_MODULE' not in settings:
-        settings = f'DJANGO_SETTINGS_MODULE={settings}'
+    if MONITORING_ENABLED:
+        settings = options.get('settings', '')
+        if settings and 'DJANGO_SETTINGS_MODULE' not in settings:
+            settings = f'DJANGO_SETTINGS_MODULE={settings}'
 
-    sh(f"{settings} python -W ignore manage.py updategeoip -o")
+        sh(f"{settings} python -W ignore manage.py updategeoip -o")
 
 
 @task
